requirements.txt
beautifulsoup4==4.13.3
fastapi==0.114.0
fastapi-cli==0.0.7
langchain
langchain-community
langchain-core
langchain-mistralai
langchain-openai
langchain-pinecone
langchain-tests
langchain-text-splitters
libsql-experimental
openai==1.61.1
pandas==2.2.3
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sse-starlette
youtube-transcript-api==0.6.2